Added to robot.cpp
Add a method that find the middle of white line and calculates how far off it is from the middle, returns how far off it is. "error" can be used by Daniel for his part on adjusting the motor speed. As the "Error" acts as how intense the motor speed must be adjusted.
Side note: 
If the error is positive it must turn right.
if the error is negative it must turn left.