Update robot.cpp
This revision of the code can complete the CORE and COMPLETION level maps - follows line far more closely than the initial version due to changing where the white pixels are being detected in the robot camera view. 

Fixes divide by zero issue that was crashing the bot occasionally when no white pixels were detected, and grants the power to turn around when no pixels are detected instead.

The management of the stopping of the program is a bit messy - might change back to while(1) rather than true, was just an easy way to check if my flag would work - could easily be replaced by a break command instead if that is desirable.

Numbers and method still need some tinkering to get it into a solid state - but this is a base that works at least.